+++ sudoku-back/src/main/kotlin/io/sudoku/sudoku/board/model/board.kt
@@ -1,0 +1,3 @@
+package io.sudoku.sudoku.board.model
+
+typealias Board = List<List<Int>>
