+++ sudoku-back/src/main/kotlin/io/sudoku/sudoku/board/service/BoardService.kt
@@ -1,0 +1,8 @@
+package io.sudoku.sudoku.board.service
+
+import io.sudoku.sudoku.board.model.Board
+
+interface BoardService {
+
+    fun createRandomMap(seed: Long?, showCount: Int):Board;
+}
